+++ CHANGELOG.md
@@ -9,6 +9,11 @@ We do try to tell you about such changes in this `CHANGELOG.md`
 and we fix bugs as fast as we can.
 
 <a id="0.5.2"></a>
+## 0.5.3 (2018-01-06)
+Can make use of `HttpParams` which yields a `request.urlWithParams`.
+Added supporting `HeroService.searchHeroes(term: string)` and test.
+
+<a id="0.5.2"></a>
 ## 0.5.2 (2017-12-10)
 No longer modify the request data coming from client. Fixes #164